+++ Assets/Scripts/UI/InGameUI/UIElements/UIClassData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0c545ed5211ee2140889ba6680acd104
+timeCreated: 1492831087
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
